Player class now has Sprite. Tidy Source tree
